backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agriwater-dashboard"
version = "0.2.0"
description = "Optimisation energetique pour l'irrigation agricole au Sahel"
requires-python = ">=3.10"
dependencies = [
    "pulp>=2.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
include = ["agriwater*", "dashboard*"]
